feat: move GET methods to /feature
